Rename old jsx network-map file into .tsx
